Move função `sha1` de `utils/files` para `extdeps/minio`
